added models file/updated routes
