+++ src/models/index.js
@@ -1,0 +1,28 @@
+let users = {
+  1: {
+    id: '1',
+    username: 'Conrad P B',
+  },
+  2: {
+    id: '2',
+    username: 'Ray Davis',
+  },
+};
+  
+let tasks = {
+  1: {
+    id: '1',
+    text: 'Shopping',
+    userId: '1',
+  },
+  2: {
+    id: '2',
+    text: 'Swimming',
+    userId: '2',
+  },
+};
+
+export default {
+  users,
+  tasks
+};
